Removes unnecessary file extension detection
Vim was instructed to detect SystemVerilog and Verilog extensions as
'verilog' and 'verilog_systemverilog' filetypes but it breaks overrides
in local vimrc files based on FileType auto commands.  The
verilog_systemverilog bundle correctly detected the normal extension
types anyway.

Also moves the pathogen ordering to be correct with syntax and filetype
after infect.